+++ imports/ui/validators/MissedBlocksTable.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
-import { CardHeader, CardBody, Card, Table, Row, Col, Nav, NavItem, NavLink, Spinner, Input } from 'reactstrap';
+import { CardBody, Card, Table, Spinner, Input } from 'reactstrap';
 import moment from 'moment';
 import { Meteor } from 'meteor/meteor';
 import numbro from 'numbro';
@@ -9,16 +9,11 @@ import PChart from '../components/Chart.jsx';
 import Account from '../components/Account.jsx';
 import { InfoIcon } from '../components/Icons.jsx'
 
+const DOWNTIMECHUCK = 4;
 const T = i18n.createComponent();
 const displayTime = (time) => {
     return moment.utc(time).format("D MMM YYYY, h:mm:ssa");
 }
-const displayTimeRange = (time) => {
-    let startTime = time.format("h:mm");
-    let endTime = time.clone().add(BATCHSIZE, 'minute').format("h:mm");
-    return `from ${startTime} to ${endTime} on ${time.format("D MMM YYYY")}`;
-}
-let DOWNTIMECHUCK = 4;
 
 const groupData = (missedStats, missedRecords, target) => {
     let validatorsMap = {};
@@ -73,7 +68,7 @@ const aggregateData = (missedRecords) => {
     return aggregatedMissedRecords;
 }
 
-class MissedBlocksTable extends Component{
+export default class MissedBlocksTable extends Component{
     constructor(props){
         super(props);
 
@@ -143,7 +138,7 @@ class MissedBlocksTable extends Component{
                 <th>Block Time</th>
                 <th>Missed Count</th>
                 <th>Missed Ratio<InfoIcon tooltipText='Missed ratio at the time of the block'/></th>
-                <th>Signed Ratio<InfoIcon tooltipText='Number of voted validators to active validators'/></th>
+                <th>Signed Ratio<InfoIcon tooltipText='Number of voted validators out of all active validators'/></th>
             </tr></thead>
             <tbody>
                 {aggregateData(data).map((record, index) => this.renderRow(record, index, grouped=grouped))}
@@ -181,203 +176,10 @@ class MissedBlocksTable extends Component{
 
     render() {
         return <Card className="missed-records-table-card">
-            <CardHeader></CardHeader>
             <CardBody>
                 <div className="float-right"> <Input type="checkbox" onClick={this.toggleGroupByValidators}/> Group By Validators</div>
                 {this.state.groupByValidators?this.renderGroupedTable():this.renderTable(this.props.missedRecords)}
             </CardBody>
         </Card>
-
     }
-}
-const BATCHSIZE = 15;
-const daysOfWeek = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat']
-class TimeDistubtionChart extends Component{
-    populateChartData() {
-        let timeline = [];
-        let breakdown = [];
-        let i;
-        for (i = 0;i < (7 * 24); i ++)
-            breakdown.push({
-                hour: i % 24,
-                day: daysOfWeek[Math.floor(i / 24)],
-                count: 0
-            });
-        let prevBatch = null;
-        this.props.missedRecords.forEach((record) => {
-            let time = moment.utc(record.time)
-            breakdown[time.day() * 24 + time.hour()].count += 1;
-            if (prevBatch && time.diff(prevBatch) >= 0) {
-                timeline[timeline.length - 1].y = timeline[timeline.length - 1].y + 1
-            } else {
-                prevBatch = time.minutes(Math.floor(time.minute()/BATCHSIZE) * BATCHSIZE).seconds(0);
-                timeline.push({
-                    x: prevBatch,
-                    y: 1
-                })
-            }
-        })
-
-
-        return {
-            timeline,
-            breakdown
-        }
-    }
-
-    populateTimelineChart(timeline) {
-        let layout = [['yAxis','barPlot'],[null, 'xAxis']];
-        let scales = [{
-            scaleId: 'xScale',
-            type: 'Time'
-        }, {
-            scaleId: 'yScale',
-            type: 'Linear'
-        }];
-        let datasets = [{
-            datasetId: 'timeline',
-            data: timeline
-        }];
-        let components = {
-            plots: [{
-                plotId: 'barPlot',
-                type: 'Bar',
-                x: {
-                    value: (d, i, ds) => d.x.toDate(),
-                    scale: 'xScale'
-                },
-                y: {
-                    value: (d, i, ds) => d.y,
-                    scale: 'yScale'
-                },
-                datasets: ['timeline'],
-                interactions: {
-                    PanZoom: {
-                        xScales: ['xScale'],
-                        yScales: ['yScale']
-                    }
-                },
-                tooltip: (c, p, data, ds) => `missed ${data.y} blocks ${displayTimeRange(data.x)}`
-            }],
-            axes: [{
-                axisId: 'xAxis',
-                type: 'Time',
-                scale: 'xScale',
-                orientation: 'bottom',
-                interactions: {
-                    PanZoom: {
-                        xScales: ['xScale']
-                    }
-                }
-            },{
-                axisId: 'yAxis',
-                type: 'Numeric',
-                scale: 'yScale',
-                orientation: 'left',
-                interactions: {
-                    PanZoom: {
-                        yScales: ['yScale']
-                    }
-                }
-            }],
-        };
-        let config = {
-            height:'300px',
-            width: '100%',
-            margin: 'auto'
-        }
-        return {layout, datasets, scales, components, config};
-    }
-
-    populateBreakDownChart(breakdown) {
-        let layout = [
-            ['yAxis','heatMap'],
-            [null, 'xAxis'],
-            [null, 'colorLegend']
-        ];
-        let scales = [{
-            scaleId: 'xScale',
-            type: 'Category'
-        }, {
-            scaleId: 'yScale',
-            type: 'Category'
-        }, {
-            scaleId: 'colorScale',
-            type: 'InterpolatedColor'
-        }];
-        let datasets = [{
-            datasetId: 'breakdown',
-            data: breakdown
-        }];
-        let components = {
-            plots: [{
-                plotId: 'heatMap',
-                type: 'Rectangle',
-                x: {
-                    value: (d, i, ds) => d.hour,
-                    scale: 'xScale'
-                },
-                y: {
-                    value: (d, i, ds) => d.day,
-                    scale: 'yScale'
-                },
-                attrs: [{
-                    attr: 'fill',
-                    value: (d) => d.count,
-                    scale: 'colorScale'
-                }, {
-                    attr: 'stroke',
-                    value: 'rgba(200, 200, 200, 0.3)'
-                }],
-                datasets: ['breakdown'],
-                tooltip: (c, p, data, ds) => `missed ${data.count} blocks on ${data.day} at ${data.hour}`
-            }],
-            axes: [{
-                axisId: 'xAxis',
-                type: 'Category',
-                scale: 'xScale',
-                orientation: 'bottom',
-            },{
-                axisId: 'yAxis',
-                type: 'Category',
-                scale: 'yScale',
-                orientation: 'left',
-                }
-            ],
-            legends: [{
-                legendId: 'colorLegend',
-                type: 'InterpolatedColor',
-                plotIds: ['heatMap'],
-                colorScaleId: 'colorScale',
-            }],
-        };
-        let config = {
-            height:'300px',
-            width: '100%',
-            margin: 'auto'
-        }
-        return {layout, datasets, scales, components, config};
-    }
-    render () {
-        let data = this.populateChartData();
-        return [
-            <Card key='timeilne'>
-                <CardHeader>History Missed Blocks</CardHeader>
-                <CardBody>
-                    <PChart {...this.populateTimelineChart(data.timeline)} />
-                </CardBody>
-            </Card>,
-            <Card key='breakdown'>
-                <CardHeader>Missed Blocks By Time of Day</CardHeader>
-                <CardBody>
-                    <PChart {...this.populateBreakDownChart(data.breakdown)} />
-                </CardBody>
-            </Card>
-        ]
-    }
-}
-
-export {
-    MissedBlocksTable,
-    TimeDistubtionChart
 }
